Support dc:creator and content:encoded elements
Adding support by adding the appropriate namespace handlers

<dc:creator>

The <author> element is spec'ed to be the author's email address. In practice,
this is undesirable so many feeds use the <dc:creator> element which is just
the author's name instead. Technically, an item should not have both an
<author> and a <dc:creator> element. If it does, last value will overwrite and
win.

<content:encoded>

This element is the RSS equivalent to Atom's content. Some feeds provide the
full text of the post. In RSS, there is an optional element called
content:encoded that roughly parallels Atom's content element, except that Atom
does not prescribe HTML encoding of the text. This adds support for the
content:encoded element.